requirements.txt
streamlit
google-generativeai
streamlit-lottie
fpdf
python-dotenv
requests
